+++ changelog.md
@@ -1,3 +1,11 @@
+### 0.0.69
+[New helper method](eanalytics_api_py/jupyter/jupyter_helper.py)
+Add a html button to trigger cell visibility according ot the given tag code
+"""
+from  eanalytics_api_py import jupyter_helper
+jupyter_helper.hide_cells(tag=["code"])
+"""
+
 ### 0.0.68
 [get_view_id_name_map](eanalytics_api_py/conn.py) Added default view_id=0